[hotfix] 리뷰 삭제 버그 수정

+++ server-api/src/main/java/com/depromeet/domains/store/service/StoreService.java
@@ -422,10 +422,11 @@ public class StoreService {
 			storeRepository.delete(store);
 		}
 
-		Integer lastIdx = imageUrl.lastIndexOf("/") + 1;
-		String fileName = imageUrl.substring(lastIdx);
-		s3Service.deleteFile(fileName);
-
+		if (imageUrl != null) {
+			Integer lastIdx = imageUrl.lastIndexOf("/") + 1;
+			String fileName = imageUrl.substring(lastIdx);
+			s3Service.deleteFile(fileName);
+		}
 	}
 
 	private void processReviewForDuplicateMostVisitor(StoreMeta storeMeta, boolean hasVisitedThreeOrMore) {
